requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
alembic==1.15.2
pydantic==2.12.5
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
httpx==0.28.1
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
